implemented the display taks method in MainWindow
i wrote a method to count the amount of taks under each priority while also counting the total ammount of task and displaying it on the stage.
i still need to write test methods and write another class to handle the counting for mainwindow
need to also write a couple more try catches a couple buttons are running erros i didnt realize before.