worked on optionsWindow viewModel
worked on OptionsWindowViewModel and did some work on creating the window for it. i added all the required proeperties and made getters and a add and remove mmethod for he comics in the viewmodel. i still need to write method specs and write test and finish the rest of part two.